requirements.txt
pyshorteners
snownlp
line-bot-sdk
Flask
gunicorn